requirements.txt
streamlit>=1.28
pandas>=1.5
plotly>=5.18
google-generativeai